first version of api wetravel

+++ app/controllers/comments_controller.rb
@@ -1,25 +1,30 @@
 class CommentsController < ApplicationController
+
+  include Pundit 
   
   def create
-
     comment = Comment.new(comment_params)
     comment.user = current_user
-    if comment.save
+    
       if params[:post_id]
         post = Post.find(params[:post_id])
         post.comments.push(comment)
       end
       if params[:comment_id]
-        comment = Comment.find(params[:comment_id])
-      render json: comment
+        commentRecep = Comment.find(params[:comment_id])
+        commentRecep.comments.push(comment)
       end
+    if comment.save
+      render json: comment.as_json(include: [:likes, :comments])
     else
-      render jsom: comment.errors
+      render json: comment.errors
     end
   end
 
   def destroy
+    
     comment = Comment.find(params[:id])
+    authorize comment
     if comment.destroy
       render json: {message: "ok"}
     else
@@ -28,6 +33,6 @@ class CommentsController < ApplicationController
   end
 
   def comment_params
-    params.permit(:body)
+    params.required(:comment).permit(:body)
   end
 end
