+++ app/controllers/likes_controller.rb
@@ -1,8 +1,10 @@
 class LikesController < ApplicationController
+
+  include Pundit 
+
   def create
     like = Like.new
     like.user = current_user
-    like.save
     if params[:post_id]
       post = Post.find(params[:post_id])
       post.likes.push(like)
@@ -11,11 +13,17 @@ class LikesController < ApplicationController
       comment = Comment.find(params[:comment_id])
       comment.likes.push(like)
     end
-    render json: like
+    if like.save
+      render json: like
+    else
+      render json: like.errors
+    end
+
   end
 
   def destroy
     like = Like.find(params[:id])
+    authorize like
     if like.destroy
       render json: {message: "ok"}
     else
