+++ app/controllers/posts_controller.rb
@@ -1,34 +1,41 @@
 class PostsController < ApplicationController
-  before_action :current_post , only: %i[show , edit , destroy] 
+  before_action :current_post , only: %i[show , update , destroy]
+  skip_before_action :require_login, only: %i[index , show]
+  include Pundit
   def index
-    @posts = Post.all
-    render json: @posts
+    posts = Post.all
+    render json: posts.map{|post| post.as_json(methods: :service_url)}
   end
 
   def show
-    render json: current_post
+    render json: current_post.as_json(methods: :service_url, include: [:likes, :comments ])
   end
     
-  def create
+  def create 
+    
     post = Post.new(post_params)
+    department = Department.find(params[:department_id])
+    post.department = department
     post.user = current_user
     if post.save
-      render json: post
+      render json: post.as_json(methods: :service_url, include: [:likes, :comments ])
     else
       render json: post.errors
     end
 
   end
    
-  def edit
+  def update
+    authorize current_post
     if(current_post.update(post_params))
-      render json: current_post
+      render json: current_post.as_json(include: [:likes, :comments])
     else
       render json: current_post.errors
     end
   end
 
   def destroy
+    authorize current_post
    if current_post.destroy
     render json: {message: "ok"}
    else
@@ -40,7 +47,7 @@ class PostsController < ApplicationController
   private 
 
   def post_params
-    params.required(:post).permit(:title, :body, :images)
+    params.required(:post).permit(:title, :body, :images )
   end
 
   def current_post
